Refactor UI changes
Made dialogs all pop correctly.  Moved some classes out of DAL and into BLL.  Made dialogs and frames have consistant names and methods.  Removed unnecessary variables and clarified others.
