+++ TrackIt/src/trackit/AnItem.java
@@ -1,7 +1,9 @@
-package trackit.DAL;
+package trackit;
 
 import java.sql.*;
 import trackit.*;
+import trackit.DAL.SQLHelper;
+import trackit.DAL.SQLHelperItem;
 
 /**
  * DAL Layer: Handles all aspects of a single Item.
