+++ TrackIt/src/trackit/DAL/SQLHelperOrderItem.java
@@ -1,5 +1,6 @@
 package trackit.DAL;
 
+import trackit.AnOrderItem;
 import java.sql.*;
 import java.util.*;
