+++ TrackIt/src/trackit/UI/OrderItemDetailsDialog.java
@@ -4,19 +4,24 @@ import java.awt.*;
 import java.awt.event.*;
 import javax.swing.*;
 import trackit.*;
-import trackit.DAL.AnOrderItem;
 
 /**
  * UI Layer: Handles all aspects of the Add Item to Order and Edit Order Item
  * dialog.
  */
 public class OrderItemDetailsDialog
-        extends ItemDetailsDialog<AnOrderItem> {
+        extends JDialog {
+
     // <editor-fold defaultstate="collapsed" desc="Constants">
+    /**
+     * The name of the dialog.
+     */
+    public static final String WINDOW_NAME = "Order Item Details";
 
     // </editor-fold>
     // <editor-fold defaultstate="expanded" desc="Private Fields">
     private final boolean isCreateMode;
+    private final AnOrderItem anOrderItem;
 
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Components">
@@ -28,13 +33,23 @@ public class OrderItemDetailsDialog
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Constructors">
     /**
+     * Creates a new dialog for working with a single Order Item.
      *
-     * @param useCreateMode
+     * @param useCreateMode True = use for a new Order Item; False = use to edit
+     * the specified Order Item.
+     * @param anOrderItem The Order Item to be edited. This value is ignored if
+     * useCreateMode is true.
      */
-    public OrderItemDetailsDialog(boolean useCreateMode) {
-        super("Order Item Details", new AnOrderItem());
-
+    public OrderItemDetailsDialog(boolean useCreateMode, AnOrderItem anOrderItem) {
         this.isCreateMode = useCreateMode;
+        if (this.isCreateMode) {
+            this.anOrderItem = null;
+        } else if (anOrderItem == null) {
+            throw new IllegalArgumentException("When 'useCreateMode' = true, then a non-null anOrderItem must be provided.");
+        } else {
+            this.anOrderItem = anOrderItem;
+        }
+
         this.initializeComponents();
     }
 
@@ -54,7 +69,7 @@ public class OrderItemDetailsDialog
         this.setLocationRelativeTo(null);
         this.setResizable(false);
         this.setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
-        //this.addWindowListener(new SupplierDetailsUI.CloseQuery());
+        this.addWindowListener(new CloseQuery());
         this.setVisible(true);
         this.getRootPane().setDefaultButton(btnOK);
 
@@ -95,15 +110,14 @@ public class OrderItemDetailsDialog
         submitBx.add(btnOK);
 
         this.btnOK.addActionListener((ActionEvent e) -> {
-            this.dispose();
+            saveAction();
         });
 
         btnCancel = new JButton("Cancel");
         submitBx.add(btnCancel);
 
         this.btnCancel.addActionListener((ActionEvent e) -> {
-            //TODO:  close window and return to prior window.
-            this.dispose();
+            cancelAction();
         });
 
         combine = Box.createVerticalBox();
@@ -114,40 +128,66 @@ public class OrderItemDetailsDialog
 
         pnlCenter.add(combine);
 
-        super.finalizeComponents();
+        //Finalizations
+        pack();
     }
 
     /**
+     * Handles the save action. If any errors, then display error message
+     * instead.
      *
      */
-    @Override
-    protected void actionSave() {
-        /*if (!bal.save()) {
-            //TODO:  display bal.getErrorMessage();
-        }*/
+    private void saveAction() {
+        JOptionPane.showMessageDialog(null, "Successfully Updated");
+        //TODO:  implement save.
+        /*if (successfullySaved) {
+                this.dispose();
+            } else {
+               //TODO:  catch errors and display them.  Do not exit dialog if an error occurs.
+            }*/
+        this.dispose();
     }
 
     /**
+     * Handles the cancel action. If any errors, then display error message
+     * instead.
      *
      */
-    @Override
-    protected void actionCancel() {
+    private void cancelAction() {
+        JOptionPane.showMessageDialog(null, "Changed Cancelled");
         //TODO:  close window and return to prior window.
+        this.dispose();
     }
 
-    /**
-     *
-     */
-    @Override
-    protected void actionDelete() {
-        /*if (bal.remove()) {
-            //TODO:  close window and return to prior window.
-        } else {
-            //TODO:  display bal.getErrorMessage() and stay on this window.
-        }*/
-    }
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Public Methods">
+    /**
+     * Displays the frame.
+     */
+    public void display() {
+        System.out.println(String.format("Displaying %s...", WINDOW_NAME));
+        setVisible(true);
+    }
 
+    // </editor-fold>
+    // <editor-fold defaultstate="collapsed" desc="SubClasses">
+    /**
+     * Handles all aspects of closing the program.
+     */
+    private class CloseQuery extends WindowAdapter {
+
+        @Override
+        public void windowClosing(WindowEvent e) {
+            JDialog frame = OrderItemDetailsDialog.this;
+            int result = JOptionPane.showConfirmDialog(frame,
+                    "Do you want to save?", "Close Query",
+                    JOptionPane.YES_NO_OPTION);
+            if (result == JOptionPane.YES_OPTION) {
+                saveAction();
+            } else {
+                cancelAction();
+            }
+        }
+    }
     // </editor-fold>
 }
