+++ TrackIt/src/trackit/UI/SupplierDetailsDialog.java
@@ -4,7 +4,6 @@ import java.awt.*;
 import java.awt.event.*;
 import javax.swing.*;
 import trackit.*;
-import trackit.DAL.ASupplier;
 
 /**
  * UI Layer: Handles all aspects of the Create ASupplier and Edit ASupplier
@@ -13,12 +12,17 @@ import trackit.DAL.ASupplier;
  * @author Douglas
  */
 public class SupplierDetailsDialog
-        extends JFrame {
+        extends JDialog {
     // <editor-fold defaultstate="collapsed" desc="Constants">
 
-    private static final String WINDOW_NAME = "Supplier Details";
-    private final ASupplier bll;
+    /**
+     * The name of the dialog.
+     */
+    public static final String WINDOW_NAME = "Supplier Details";
+    // </editor-fold>
+    // <editor-fold defaultstate="collapsed" desc="Private Fields">
     private final boolean isCreateMode;
+    private final ASupplier aSupplier;
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Components">
     JPanel pnlCenter;
@@ -29,17 +33,28 @@ public class SupplierDetailsDialog
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Constructors">
     /**
+     * Creates a new dialog for working with a single Supplier.
      *
-     * @param useCreateMode
+     * @param useCreateMode True = use for a new Supplier; False = use to edit
+     * the specified Supplier.
+     * @param aSupplier The Supplier to be edited. This value is ignored if
+     * useCreateMode is true.
      */
-    public SupplierDetailsDialog(boolean useCreateMode) {
-        this.bll = new ASupplier();
+    public SupplierDetailsDialog(boolean useCreateMode, ASupplier aSupplier) {
         this.isCreateMode = useCreateMode;
+        if (this.isCreateMode) {
+            this.aSupplier = null;
+        } else if (aSupplier == null) {
+            throw new IllegalArgumentException("When 'useCreateMode' = true, then a non-null aSupplier must be provided.");
+        } else {
+            this.aSupplier = aSupplier;
+        }
+
         initializeComponents();
     }
+
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Private Methods">
-
     /**
      * Sets up all components used in this frame.
      */
@@ -78,20 +93,14 @@ public class SupplierDetailsDialog
         submitBx.add(btnOK);
 
         this.btnOK.addActionListener((ActionEvent e) -> {
-            this.dispose();
-
-            /* if (!bll.save()) {
-                //TODO:  display bal.getErrorMessage();
-
-            }*/
+            saveAction();
         });
 
         btnCancel = new JButton("Cancel");
         submitBx.add(btnCancel);
 
         this.btnCancel.addActionListener((ActionEvent e) -> {
-            this.dispose();
-            //TODO:  close window and return to prior window.
+            cancelAction();
         });
         combine = Box.createVerticalBox();
         combine.add(nameBx);
@@ -102,6 +111,33 @@ public class SupplierDetailsDialog
 
         //Finalizations
         pack();
+    }
+
+    /**
+     * Handles the save action. If any errors, then display error message
+     * instead.
+     *
+     */
+    private void saveAction() {
+        JOptionPane.showMessageDialog(null, "Successfully Updated");
+        //TODO:  implement save.
+        /*if (successfullySaved) {
+                this.dispose();
+            } else {
+               //TODO:  catch errors and display them.  Do not exit dialog if an error occurs.
+            }*/
+        this.dispose();
+    }
+
+    /**
+     * Handles the cancel action. If any errors, then display error message
+     * instead.
+     *
+     */
+    private void cancelAction() {
+        JOptionPane.showMessageDialog(null, "Changed Cancelled");
+        //TODO:  close window and return to prior window.
+        this.dispose();
     }
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Public Methods">
@@ -114,13 +150,6 @@ public class SupplierDetailsDialog
         setVisible(true);
     }
 
-    /**
-     * close window
-     */
-    public void closeWindow() {
-        this.setVisible(false);
-    }
-
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="SubClasses">
     /**
@@ -130,17 +159,14 @@ public class SupplierDetailsDialog
 
         @Override
         public void windowClosing(WindowEvent e) {
-            JFrame frame = (JFrame) e.getSource();
+            JDialog frame = SupplierDetailsDialog.this;
             int result = JOptionPane.showConfirmDialog(frame,
                     "Do you want to save?", "Close Query",
                     JOptionPane.YES_NO_OPTION);
             if (result == JOptionPane.YES_OPTION) {
-                //TODO
-                JOptionPane.showMessageDialog(null, "Successfully Updated");
-                closeWindow();
+                saveAction();
             } else {
-                //TODO
-                closeWindow();
+                cancelAction();
             }
         }
     }
