+++ TrackIt/src/trackit/UI/SuppliersPanel.java
@@ -4,7 +4,7 @@ import java.awt.*;
 import java.awt.event.*;
 import java.util.*;
 import javax.swing.*;
-import trackit.DAL.ASupplier;
+import trackit.ASupplier;
 
 /**
  * UI Layer: Handles all aspects of the Suppliers panel.
@@ -19,16 +19,15 @@ public class SuppliersPanel
      * The name of the panel.
      */
     public static final String TAB_NAME = "Suppliers";
+    private static final String[] TABLE_LABELS = {"Supplier", "Web Address"};
     // </editor-fold>
     // <editor-fold defaultstate="expanded" desc="Private Fields">
-    private final ArrayList<ASupplier> suppliers;
+    private final ArrayList<ASupplier> suppliers = new ArrayList<>();
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Components">
     JButton btnCreate, btnRemove, btnEdit;
-    String[] suppliersLabel = {"Supplier", "Web Address"};
-    JTable suppliersTable;
+    JTable mainTable;
     SupplierDetailsDialog details;
-    int selectedRow;
 
     // </editor-fold>
     // <editor-fold defaultstate="collapsed" desc="Constructors">
@@ -36,15 +35,20 @@ public class SuppliersPanel
      * Supplier UI
      */
     public SuppliersPanel() {
-        this.suppliers = new ArrayList<>();
+        initializeComponents();
+    }
+    // </editor-fold>
+    // <editor-fold defaultstate="collapsed" desc="Private Methods">
+
+    private void initializeComponents() {
         setLayout(new BorderLayout());
 
         //add data to suppliers arraylist 
         Object[][] suppliersTestData = {{"Amazon", "http://www.amazon.com"}, {"Walmart", "http://www.walmart.com"}, {"Ebay", "http://www.ebay.com"}};
-        suppliersTable = new JTable(suppliersTestData, suppliersLabel);
-        JScrollPane suppliersScrollPane = new JScrollPane(suppliersTable);
-        suppliersTable.setFillsViewportHeight(true);
-        suppliersTable.setDefaultEditor(Object.class, null);
+        mainTable = new JTable(suppliersTestData, TABLE_LABELS);
+        JScrollPane suppliersScrollPane = new JScrollPane(mainTable);
+        mainTable.setFillsViewportHeight(true);
+        mainTable.setDefaultEditor(Object.class, null);
 
         add(suppliersScrollPane, BorderLayout.CENTER);
 
@@ -52,33 +56,43 @@ public class SuppliersPanel
 
         btnCreate = new JButton("Create");
         btnCreate.addActionListener((ActionEvent e) -> {
-            System.out.print("create supply");
-            details = new SupplierDetailsDialog(true);
+            //System.out.print("create supply");
+            SupplierDetailsDialog dlgCreate = new SupplierDetailsDialog(true, null);
+            dlgCreate.display();
         });
 
         btnEdit = new JButton("Edit");
         btnEdit.addActionListener((ActionEvent e) -> {
-            System.out.print("Edit supply");
-            //if list item selected edit item else select item
-            selectedRow = suppliersTable.getSelectedRow();
+            //System.out.print("Edit supply");
+            //If list item selected then edit item else select item.
+            int selectedRow = mainTable.getSelectedRow();
             if (selectedRow < 0) {
-                JOptionPane.showMessageDialog(null, "Select item to edit");
+                JOptionPane.showMessageDialog(this, "Select item to edit");
             } else {
-                details = new SupplierDetailsDialog(false);
-                //TODO: enter item info of selected item
+                ASupplier aSupplier = new ASupplier();
+                //TODO: Set aSupplier to the value of selectedRow.
+                SupplierDetailsDialog dlgEdit = new SupplierDetailsDialog(false, aSupplier);
+                dlgEdit.display();
             }
         });
 
         btnRemove = new JButton("Remove");
         btnRemove.addActionListener((ActionEvent e) -> {
-            System.out.print("remove supply");
-            selectedRow = suppliersTable.getSelectedRow();
+            int selectedRow = this.mainTable.getSelectedRow();
             if (selectedRow < 0) {
                 JOptionPane.showMessageDialog(null, "Select item to remove");
             } else {
                 //TODO: remove item from db
                 JOptionPane.showMessageDialog(null, "Item removed");
             }
+            //TODO: surround below in a for loop
+            /*
+            if (bll.remove()) {
+                //TODO:  close window and return to prior window.
+            } else {
+                //TODO:  display bll.getErrorMessage() and stay on this window.
+            }
+             */
         });
 
         btmSup.add(btnCreate);
